Triangulation

+++ Shape/Circle.cs
@@ -1,16 +1,18 @@
+using LSPainter.Maths;
+
 namespace LSPainter.Shapes
 {
     public class Circle : Shape
     {
         public Vector Origin { get; }
-        public float Radius { get; }
+        public double Radius { get; }
 
-        public Circle(Vector origin, float radius)
+        public Circle(Vector origin, double radius)
         {
             Origin = origin;
             Radius = radius;
 
-            Area = (float)Math.PI * Radius * Radius;
+            Area = Math.PI * Radius * Radius;
 
             BoundingBox = new Rectangle(Origin.X - Radius, Origin.Y - Radius, Radius * 2, Radius * 2);
         }
